[ASTL-355] Use CLOCK_MONOTONIC_RAW for processed samples and add clock correlation for raw samples (#403)
## Description

This PR introduces per-operation clock correlation so raw sample
timestamps from different collector-native clock domains can be
normalized to a common CLOCK_MONOTONIC_RAW timeline before being
processed/output.

Changes:

Add a per-operation clock-correlation snapshot API to
collectors/collector-manager and invoke it at collection start.
Store correlations in MetricManager and normalize raw sample timestamps
during ProcessRawSamples.
Introduce ProcessedSampleTimestamp (based on CLOCK_MONOTONIC_RAW) and
update outputs + tests to use it.

Fixes # ASTL-355

## Customer Impact

Users will receive samples in a uniform timestamp based on
CLOCK_MONOTONIC_RAW

## Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Testing, Verification, or CICD change
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Interface change (fix or feature that would cause the user facing
interface to change)
- [ ] This change requires a documentation update
- [ ] This changes requires a version update

## How Has This Been Tested?

Please describe the tests that you ran to verify your changes. Provide
instructions so we can reproduce. Please also list any relevant details
for your test configuration

- [x] unit testing
- [x] CI/CD

## Test Configuration

OS name: MacOS, Linux
OS version: Ubuntu 24.04.3 LTS, MacOS 15.7.4
Kernel version: Linux asct-astra2 6.16.3, Darwin Kernel Version 24.6.0
Hardware name: Mac, Astra Linux

## Checklist

- [x] My code follows the style guidelines of this project
- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [x] I have added tests that prove my fix is effective or that my
feature works
- [x] New and existing unit tests pass locally with my changes
- [x] Any dependent changes have been merged and published in downstream
modules
- [x] I have checked my code and corrected any misspellings

---------


Ossmosis-Source-Commit: ea9cd916871574fe6125381ebe489d1c6ec22bd2

+++ src/astl_telemetry.cpp
@@ -1893,7 +1893,8 @@ auto astlGetCounterSamplesOnTarget(const astl_get_counter_samples_on_target_para
   // helper lambda to convert a ProcessedSampledData into an astl_sample_t
   auto convert_to_counter_sample = [](const astl::ProcessedSampledData& processed_sample) {
     const auto union_value = processed_sample.value.ToAstlUnionValue().first;  // avoid constructing pair twice
-    return astl_sample_t{.timestamp = processed_sample.timestamp.time_since_epoch().count(), .value = union_value};
+    return astl_sample_t{.timestamp = static_cast<uint64_t>(processed_sample.timestamp.time_since_epoch().count()),
+                         .value     = union_value};
   };
 
   // samples_span is at least large enough to accomodate sample_result because of above check.
